+++ src/features/main-screen/index.tsx
@@ -1,9 +1,13 @@
-import { FC, useEffect, useState } from 'react'
-import { handleSuccess } from './audio-recorder/handler/handle-success'
-import { initWebsocket } from './audio-recorder/handler/init-websocket'
-import { LoadingSpinner } from './loading-spinner'
-import { getTranslation } from '../lib/sotra-manager'
-import { getParseDataForYoutube } from '../lib/youtube-manager'
+import { useEffect, useMemo, useRef, useState } from 'react'
+import { handleSuccess } from './components/audio-recorder/handler/handle-success'
+import { initWebsocket } from './components/audio-recorder/handler/init-websocket'
+import { LoadingSpinner } from '../../components/loading-spinner'
+import { getTranslation } from '../../lib/sotra-manager'
+import { getParseDataForYoutube } from '../../lib/youtube-manager'
+import { MicrophoneSelector } from './components/microphone-selector.tsx'
+import { RecordButtonsContainer } from './components/record-buttons-container'
+import { Box } from '@mui/material'
+import { Visualizer } from 'react-sound-visualizer'
 
 const SAMPLE_RATE = 48000
 let processor: AudioWorkletNode
@@ -13,26 +17,27 @@ let context: AudioContext
 let localeStream: MediaStream
 
 const youtubeUrl =
-  'http://upload.youtube.com/closedcaption?cid=1234-5678-9012-3456'
+  'http://upload.youtube./com/closedcaption?cid=1234-5678-9012-3456'
 let seq = 0
 
-export const AudioRecorder: FC<{}> = () => {
+export const MainScreen = () => {
   const [inputText, setInputText] = useState<string>('')
   const [translation, setTranslation] = useState<string>('')
   const [youtubeSubtitle, setYoutubeSubtitle] = useState<string>('')
-
+  const [selectedMicrophone, setSelectedMicrophone] =
+    useState<MediaDeviceInfo | null>(null)
   const [isRecording, setIsRecording] = useState<boolean>(false)
 
   const onReceiveMessage = (event: MessageEvent) => {
     if (event.data) {
       let parsed = JSON.parse(event.data)
+      console.log(parsed)
       if (
         parsed.text &&
         parsed.text !== '-- ***/whisper/ggml-model.q8_0.bin --' &&
         parsed.text !== '-- **/whisper/ggml-model.q8_0.bin --' &&
         parsed.text !== '-- */whisper/ggml-model.q8_0.bin --'
       ) {
-        const now = new Date()
         seq += 1
         const trimmedText = parsed.text.slice(2, -2).trim()
         setInputText((prev) => prev + ' ' + trimmedText)
@@ -40,7 +45,7 @@ export const AudioRecorder: FC<{}> = () => {
           const youtubeData = getParseDataForYoutube(
             seq,
             response.data.translation,
-            now,
+            new Date(parsed.start * 1000),
             youtubeUrl,
             false
           )
@@ -89,20 +94,22 @@ export const AudioRecorder: FC<{}> = () => {
               channelCount: 1,
               sampleRate: SAMPLE_RATE,
               sampleSize: 16,
+              deviceId: selectedMicrophone?.deviceId,
             },
             video: false,
           })
           .then((stream) => {
             localeStream = stream
             setIsRecording(true)
-            handleSuccess(
-              stream,
-              SAMPLE_RATE,
-              webSocket,
-              onSetNewProcessor,
-              onSetNewSource,
-              onSetNewContext
-            )
+            if (localeStream !== null)
+              handleSuccess(
+                localeStream,
+                SAMPLE_RATE,
+                webSocket,
+                onSetNewProcessor,
+                onSetNewSource,
+                onSetNewContext
+              )
           })
       } catch (error) {
         console.error('Error accessing microphone:', error)
@@ -126,6 +133,24 @@ export const AudioRecorder: FC<{}> = () => {
       setIsRecording(false)
     }
   }
+
+  const pauseRecording = () => {}
+
+  const visualizerArea = useMemo(
+    () => (
+      <Visualizer
+        audio={localeStream}
+        mode='continuous'
+        autoStart
+        strokeColor='white'
+        lineWidth='default'
+      >
+        {({ canvasRef }) => <canvas ref={canvasRef} width={500} height={100} />}
+      </Visualizer>
+    ),
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [localeStream]
+  )
 
   return (
     <div
@@ -152,22 +177,55 @@ export const AudioRecorder: FC<{}> = () => {
           <LoadingSpinner />
         </div>
       )}
-      <p>{inputText}</p>
-      <div style={{ height: 1, width: '80%', backgroundColor: 'white' }} />
-      <p>{translation}</p>
-      <button onClick={startRecording} disabled={isRecording}>
-        Start Recording
-      </button>
-      <button onClick={stopRecording} disabled={!isRecording}>
-        Stop Recording
-      </button>
+      <Box
+        sx={{
+          display: 'flex',
+          flexDirection: 'row',
+          justifyContent: 'center',
+          alignItems: 'center',
+          gap: 2,
+        }}
+      >
+        <Box
+          style={
+            selectedMicrophone === null
+              ? {}
+              : { position: 'absolute', top: 10, right: 10 }
+          }
+        >
+          <MicrophoneSelector
+            onChange={(mic) => {
+              setSelectedMicrophone(mic)
+              stopRecording()
+            }}
+          />
+        </Box>
+        {localeStream && visualizerArea}
+        <RecordButtonsContainer
+          isDisabled={{
+            record: isRecording || !selectedMicrophone,
+            pause: !isRecording,
+            stop: !isRecording,
+          }}
+          onPressRecord={startRecording}
+          onPressPause={pauseRecording}
+          onPressStop={stopRecording}
+        />
+      </Box>
 
-      <textarea
-        color='white'
-        style={{ width: '300px', minHeight: '400px' }}
-        value={youtubeSubtitle}
-        readOnly
-      />
+      {selectedMicrophone && (
+        <>
+          <p>{inputText}</p>
+          <div style={{ height: 1, width: '80%', backgroundColor: 'white' }} />
+          <p>{translation}</p>
+          <textarea
+            color='white'
+            style={{ width: '300px', minHeight: '400px' }}
+            value={youtubeSubtitle}
+            readOnly
+          />
+        </>
+      )}
     </div>
   )
 }