+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1 (COMMING SOON)
 
 - auto-scroll za wšitke pola z tekstom
+- selector mikrofonow
 
 ## 1.0.0 (RELEASED)